--- bamboo-iOS/Sources/Scene/Modal/BulletinBoardWriteModal/Protocol/WritingBulletinBoardModalProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  WritingBulletinBoardModalProtocol.swift
-//  bamboo-iOS
-//
-//  Created by Ji-hoon Ahn on 2021/10/03.
-//
-
-import Foundation
-
-protocol WriteModalDelegate : AnyObject{
-    func updateWrite()
-}
