--- bamboo-iOS/Sources/Scene/Modal/EditContentModal/Protocol/EditContentModalProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  EditContentModalProtocol.swift
-//  bamboo-iOS
-//
-//  Created by Ji-hoon Ahn on 2021/10/24.
-//
-
-import Foundation
-
-protocol EditContentModalProtocol : AnyObject{
-    func onTapClose()
-}
