--- bamboo-iOS/Sources/Scene/Modal/ManagerLoginModal/Protocol/ManagerModalDelegate.swift
@@ -1,12 +1,0 @@
-//
-//  ManagerModalDelegate.swift
-//  bamboo-iOS
-//
-//  Created by Ji-hoon Ahn on 2021/11/02.
-//
-
-import Foundation
-
-protocol ManagerModalDelegate : AnyObject {
-    func updateManagerModal()
-}
