--- bamboo-iOS/Sources/Scene/Modal/RefusalModal/Protocol/RefusalModalProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  RefusalModalProtocol.swift
-//  bamboo-iOS
-//
-//  Created by Ji-hoon Ahn on 2021/10/26.
-//
-
-import Foundation
-
-protocol RefusalModalProtocol : AnyObject{
-    func onTapClose()
-}
