--- bamboo-iOS/Sources/Scene/Modal/ReportModal/Protocol/ReportModalProtocol.swift
@@ -1,13 +1,0 @@
-//
-//  ReportModalProtocol.swift
-//  bamboo-iOS
-//
-//  Created by Ji-hoon Ahn on 2021/10/28.
-//
-
-import Foundation
-
-protocol ReportModalDelegate : AnyObject {
-    func updateReport()
-}
-
