+++ bamboo-iOS/Sources/Views/Loading/LoadingIndicator.swift
