+++ bamboo-iOS/Sources/Views/TableViewCell/Delete/DeleteTableViewCell.swift
